requirements.txt
numpy==1.19.5
opencv-contrib-python==4.6.0.66
python-dotenv==0.20.0
dnspython==2.2.1
pymongo==4.2.0